Refactor package structure for child management components; move classes to appropriate internal service and repository packages
