+++ src/main/java/com/careconnect/coreapi/childmgmt/internal/controller/AllergyController.java
@@ -1,8 +1,8 @@
-package com.careconnect.coreapi.childmgmt.internal;
+package com.careconnect.coreapi.childmgmt.internal.controller;
 
 import com.careconnect.coreapi.childmgmt.domain.Allergy;
 import com.careconnect.coreapi.childmgmt.domain.ChildAllergy;
-import com.careconnect.coreapi.childmgmt.internal.AllergyService;
+import com.careconnect.coreapi.childmgmt.internal.service.AllergyService;
 import com.careconnect.coreapi.common.response.ApiResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
