+++ src/main/java/com/careconnect/coreapi/childmgmt/internal/repository/ChildGuardianRepository.java
@@ -1,4 +1,4 @@
-package com.careconnect.coreapi.childmgmt.internal;
+package com.careconnect.coreapi.childmgmt.internal.repository;
 
 import com.careconnect.coreapi.childmgmt.domain.ChildGuardian;
 import org.springframework.data.jpa.repository.JpaRepository;
