+++ src/main/java/com/careconnect/coreapi/childmgmt/internal/service/GuardianService.java
@@ -1,7 +1,7 @@
-package com.careconnect.coreapi.childmgmt.internal;
+package com.careconnect.coreapi.childmgmt.internal.service;
 
 import com.careconnect.coreapi.childmgmt.domain.Guardian;
-import com.careconnect.coreapi.childmgmt.internal.GuardianRepository;
+import com.careconnect.coreapi.childmgmt.internal.repository.GuardianRepository;
 import com.careconnect.coreapi.common.exceptions.ResourceNotFoundException;
 import com.careconnect.coreapi.common.exceptions.ValidationException;
 import com.careconnect.coreapi.user.UserService;
