+++ src/test/java/com/careconnect/coreapi/childmgmt/repository/ChildRepositoryTest.java
@@ -2,7 +2,7 @@ package com.careconnect.coreapi.childmgmt.repository;
 
 import com.careconnect.coreapi.childmgmt.domain.Child;
 import com.careconnect.coreapi.childmgmt.domain.Guardian;
-import com.careconnect.coreapi.childmgmt.internal.ChildRepository;
+import com.careconnect.coreapi.childmgmt.internal.repository.ChildRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
