+++ src/test/java/com/careconnect/coreapi/childmgmt/service/ChildGuardianServiceTest.java
@@ -3,10 +3,10 @@ package com.careconnect.coreapi.childmgmt.service;
 import com.careconnect.coreapi.childmgmt.domain.Child;
 import com.careconnect.coreapi.childmgmt.domain.ChildGuardian;
 import com.careconnect.coreapi.childmgmt.domain.Guardian;
-import com.careconnect.coreapi.childmgmt.internal.ChildGuardianRepository;
-import com.careconnect.coreapi.childmgmt.internal.ChildGuardianService;
-import com.careconnect.coreapi.childmgmt.internal.ChildRepository;
-import com.careconnect.coreapi.childmgmt.internal.GuardianRepository;
+import com.careconnect.coreapi.childmgmt.internal.repository.ChildGuardianRepository;
+import com.careconnect.coreapi.childmgmt.internal.service.ChildGuardianService;
+import com.careconnect.coreapi.childmgmt.internal.repository.ChildRepository;
+import com.careconnect.coreapi.childmgmt.internal.repository.GuardianRepository;
 import com.careconnect.coreapi.common.exceptions.DuplicateResourceException;
 import com.careconnect.coreapi.common.exceptions.ResourceNotFoundException;
 import com.careconnect.coreapi.common.exceptions.ValidationException;
