+++ src/test/java/com/careconnect/coreapi/childmgmt/service/ChildServiceTest.java
@@ -2,9 +2,9 @@ package com.careconnect.coreapi.childmgmt.service;
 
 import com.careconnect.coreapi.childmgmt.domain.Child;
 import com.careconnect.coreapi.childmgmt.domain.Guardian;
-import com.careconnect.coreapi.childmgmt.internal.ChildRepository;
-import com.careconnect.coreapi.childmgmt.internal.ChildService;
-import com.careconnect.coreapi.childmgmt.internal.GuardianRepository;
+import com.careconnect.coreapi.childmgmt.internal.repository.ChildRepository;
+import com.careconnect.coreapi.childmgmt.internal.service.ChildService;
+import com.careconnect.coreapi.childmgmt.internal.repository.GuardianRepository;
 import com.careconnect.coreapi.common.exceptions.ResourceNotFoundException;
 import com.careconnect.coreapi.common.exceptions.ValidationException;
 import org.junit.jupiter.api.BeforeEach;
